Combining common registers
